remove Logo from components, because it caused problems with prod build
