+++ src/common/components/Sidebar/index.jsx
@@ -1,19 +1,28 @@
 import React, {Component} from 'react'
 import PropTypes from 'prop-types'
 import {NavLink} from 'react-router-dom'
-import {Menu, Icon} from 'semantic-ui-react'
-import {Logo} from 'components'
-// import _ from 'lodash'
+import {Menu, Sidebar, Icon, Image} from 'semantic-ui-react'
 import './Sidebar.scss'
 
-export default class SidebarInnerComponent extends Component {
+export default class SidebarComponent extends Component {
   static propTypes = {
+    open: PropTypes.bool,
     logout: PropTypes.func,
-    routing: PropTypes.array
+    routing: PropTypes.array,
+    isMobile: PropTypes.bool
   }
 
   render () {
-    const {logout, routing} = this.props
+    const {open, logout, routing, isMobile} = this.props
+
+    const sidebarProps = {
+      visible: open || !isMobile,
+      as: Menu,
+      vertical: true,
+      icon: 'labeled',
+      animation: 'push',
+      width: 'thin'
+    }
 
     const routes = routing.map((route, i) => {
       const {external, path, icon, name, strict, exact} = route
@@ -38,16 +47,22 @@ export default class SidebarInnerComponent extends Component {
         </Menu.Item>
       )
     })
-
+    // XXX: @Metnew 12.06.2017:
+    // it's recommended to create separate Logo component for app
+    // But I caught 130# error in production build multiple times (invalid component type)
+    // When I've used Logo component (see repo history)
+    // I still don't know what was the problem behind it.
     return (
-      <div>
-        <Logo centered />
+      <Sidebar {...sidebarProps}>
+        <div className="logo">
+          <Image src="./images/logo.png" centered height="34px" />
+        </div>
         {routes}
         <Menu.Item className="logout" onClick={logout}>
           <Icon name="sign out" />
           Logout
         </Menu.Item>
-      </div>
+      </Sidebar>
     )
   }
 }
